requirements.txt
# Core
fastapi
uvicorn
pydantic
python-dotenv
pymongo
pandas
streamlit

# Google Gemini
google-generativeai

# RAG and Embeddings
faiss-cpu
sentence-transformers

# HTTP Requests
requests

# Optional (for async performance)
httpx
